+++ Sources/ApodiniDeploy/ApodiniDeployInterfaceExporter.swift
@@ -139,7 +139,6 @@ class ApodiniDeployInterfaceExporter: InterfaceExporter {
     
     func finishedExporting(_ webService: WebServiceModel) {
         do {
-//            try exportWebServiceStructure(apodiniDeployConfiguration: self.exporterConfiguration)
             try performDeploymentRelatedActions()
         } catch {
             fatalError("Error performing deployment-related actions: \(error)")
@@ -149,7 +148,7 @@ class ApodiniDeployInterfaceExporter: InterfaceExporter {
     
     private func performDeploymentRelatedActions() throws {
         let env = ProcessInfo.processInfo.environment
-        
+        print(#function)
         // If those enironment variables do not exist, return immediatly
         guard let mode = env[WellKnownEnvironmentVariables.executionMode],
               let fileURL = env[WellKnownEnvironmentVariables.fileUrl]
@@ -196,8 +195,8 @@ class ApodiniDeployInterfaceExporter: InterfaceExporter {
             } catch {
                 throw ApodiniDeployError(message: "Unable to launch with custom config: \(error)")
             }
-            
         default:
+            try exportWebServiceStructure(apodiniDeployConfiguration: self.exporterConfiguration)
             break
         }
     }
