+++ Sources/ApodiniDeployBuildSupport/DeploymentProvider.swift
@@ -136,71 +136,6 @@ extension DeploymentProvider {
     }
     
     
-    /// Read the web service's structure, and return it encoded as a `WebServiceStructure` object
-//    public func readWebServiceStructure() throws -> WebServiceStructure {
-//        print("read ws structure")
-//        let fileManager = FileManager()
-//        let logger = Logger(label: "ApodiniDeployCLI.Localhost")
-//        
-//        let modelFileUrl = fileManager.temporaryDirectory.appendingPathComponent("AM_\(UUID().uuidString).json")
-//        guard fileManager.createFile(atPath: modelFileUrl.path, contents: nil, attributes: nil) else {
-//            throw ApodiniDeployBuildSupportError(message: "Unable to create file")
-//        }
-//        
-//        let exportWebServiceModelTask: Task
-//        
-//        switch target {
-//        case .executable(let executableUrl):
-//            exportWebServiceModelTask = Task(
-//                executableUrl: executableUrl,
-//                captureOutput: false,
-//                launchInCurrentProcessGroup: launchChildrenInCurrentProcessGroup,
-//                environment: [
-//                    WellKnownEnvironmentVariables.executionMode:
-//                        WellKnownEnvironmentVariableExecutionMode.exportWebServiceModelStructure,
-//                    WellKnownEnvironmentVariables.fileUrl: modelFileUrl.path
-//                ]
-//            )
-//        case let .spmTarget(packageUrl, productName):
-//            let swiftBin = try Self.getSwiftBinUrl()
-//            guard fileManager.directoryExists(atUrl: packageUrl) else {
-//                throw ApodiniDeployBuildSupportError(message: "Unable to find input directory")
-//            }
-//            try fileManager.setWorkingDirectory(to: packageUrl)
-//            
-//            let packageSwiftFileUrl = packageUrl.appendingPathComponent("Package.swift")
-//            guard fileManager.fileExists(atPath: packageSwiftFileUrl.path) else {
-//                throw ApodiniDeployBuildSupportError(message: "Unable to find Package.swift")
-//            }
-//            exportWebServiceModelTask = Task(
-//                executableUrl: swiftBin,
-//                arguments: [
-//                    "run",
-//                    productName
-//                ],
-//                captureOutput: false,
-//                launchInCurrentProcessGroup: launchChildrenInCurrentProcessGroup,
-//                environment: [
-//                    WellKnownEnvironmentVariables.executionMode:
-//                        WellKnownEnvironmentVariableExecutionMode.exportWebServiceModelStructure,
-//                    WellKnownEnvironmentVariables.fileUrl: modelFileUrl.path
-//                ]
-//            )
-//        }
-//        
-//        logger.notice("Invoking child process `\(exportWebServiceModelTask)`")
-//        let terminationInfo = try exportWebServiceModelTask.launchSync()
-//        guard terminationInfo.exitCode == EXIT_SUCCESS else {
-//            throw ApodiniDeployBuildSupportError(message: "Unable to generate model structure: \(terminationInfo.exitCode)")
-//        }
-//        
-//        logger.notice("model written to '\(modelFileUrl)'")
-//        
-//        let data = try Data(contentsOf: modelFileUrl, options: [])
-//        return try JSONDecoder().decode(WebServiceStructure.self, from: data)
-//    }
-    
-    
     /// Based on a `WebServiceStructure` as input (i.e. a representation of an entire web service),
     /// computes the default set of nodes within the deployed system.
     public func computeDefaultDeployedSystemNodes(
